Reorganized fishtrap module
