+++ src/java/growthcraft/fishtrap/entity/TileEntityFishTrap.java
@@ -1,4 +1,4 @@
-package growthcraft.fishtrap;
+package growthcraft.fishtrap.entity;
 
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.inventory.IInventory;
